Rename tests for cli/test_capsule.py.
Test case naming follows #3038 conventions.
